requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
networkx==3.3
pydantic==2.7.1
web3==6.19.0
neo4j==5.22.0
requests==2.32.3
pynacl==1.5.0
PyJWT==2.9.0
